correciones de reviews
