+++ src/app/components/reportesDiarios/planillaC1.component.ts
@@ -1,22 +1,20 @@
 import { Plex } from '@andes/plex';
 import { Router } from '@angular/router';
 import { Component, Input, OnInit, HostBinding, Output, EventEmitter } from '@angular/core';
-import { FormBuilder, FormGroup, Validators, FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { Server } from '@andes/shared';
 import { Auth } from '@andes/auth';
-import * as moment from 'moment';
-import { dateValidator } from '@andes/plex/src/lib/core/validator.functions';
-import { combineAll } from 'rxjs/operators';
-
-
 
 
 @Component({
-    selector: 'resumenDiarioMensual',
-    templateUrl: 'resumenDiarioMensual.html',
+    selector: 'planillaC1',
+    templateUrl: 'planillaC1.html',
+    styleUrls: [
+        'planillaC1.scss'
+    ]
 })
 
-export class ResumenDiarioMensualComponent implements OnInit {
+
+export class PlanillaC1Component implements OnInit {
 
     fecha = new Date();
