+++ src/app/components/reportesDiarios/planillaC1.scss
@@ -1,0 +1,13 @@
+
+    th,
+    td {
+        text-align: center;
+        width: 25px;
+    }
+
+
+
+    .table th,
+    .table td {
+        padding: 0.2rem;
+    }
