+++ src/app/components/reportesDiarios/resumenDiarioMensual.scss
@@ -1,0 +1,11 @@
+
+    th,td{
+        text-align: center;
+        width: 25px;
+    }
+
+
+
+    .table th, .table td {
+        padding: 0.2rem;
+    }
