--- src/app/components/reportesSJ/encabezadoReportesSJ.html
@@ -1,102 +1,0 @@
-<!-- <style>
-    .content{
-        height: 1600px;
-    }
-</style> -->
-
-<form class="plex-layout" #form="ngForm">
-    <section>
-        <div class="row">
-            <div class="col">
-                <plex-box>
-                    <header>
-                        <div class="clearfix">
-                            <div class="page-title float-left">Reportes San Juan</div>
-                        </div>
-
-                        <div class="row">
-                            <br>
-                            <!-- Select organizacion -->
-                            <div class="col-4">
-                                <plex-select [(ngModel)]="organizacion" name="organizacion" (getData)="loadOrganizacion($event)"
-                                    label="Organización" placeholder="Seleccione la organización" (change)="refreshSelection($event,'organizacion')"
-                                    labelField="nombre" [required]="true">
-                                </plex-select>
-                            </div>
-                            <br>
-
-                            <!-- Select Reporte -->
-                            <div class="col-4">
-                                <plex-select [(ngModel)]="tipoReportes" name="tipoReportes" [data]="opcionesReportes"
-                                    label-field="Tipo de reportes" label="Seleccione reporte" placeholder="Seleccione reporte"
-                                    [required]="true" (change)="onChangeTipoReportes($event);refreshSelection($event,'tipoReportes')"
-                                    [readonly]="organizacion === null">
-                                </plex-select>
-                            </div>
-                        </div>
-
-                        <!-- Parametros ResumenDiarioMensual -->
-                        <div class="row" *ngIf="organizacion!==null && tipoReportes !==null && tipoReportes.nombre ==='ResumenDiarioMensual'">
-
-                            <div class="col-2">
-                                <!-- Select Mes -->
-                                <plex-select [(ngModel)]="mes" name="mes" [data]="opcionesMes" label="Mes" (change)="refreshSelection($event,'mes')"
-                                    placeholder="Mes" [required]="true">
-                                </plex-select>
-                            </div>
-
-                            <div class="col-2">
-                                <!-- Select Año -->
-                                <plex-select [(ngModel)]="anio" name="año" [data]="opcionesAnio" label="Año" (change)="refreshSelection($event,'anio')"
-                                    placeholder="Año" [required]="true">
-                                </plex-select>
-                            </div>
-
-                            <div class="col-4">
-                                <!-- Select Unidad Operativa -->
-                                <plex-select [(ngModel)]="unidadOperativa" name="unidadOperativa" [data]="opcionesUnidadOperativa"
-                                    label="Unidad Operativa" placeholder="Seleccione la unidad operativa" labelField="nombre"
-                                    (change)="refreshSelection($event,'unidadOperativa')" [required]="true" [readonly]="(organizacion === null) || (mes===null) || (anio===null)">
-                                </plex-select>
-                            </div>
-                        </div>
-
-
-
-                        <!-- Parametros ResumenDiarioMensual -->
-                        <div class="row" *ngIf="organizacion!==null && tipoReportes !==null && tipoReportes.nombre ==='PlanillaC1'">
-
-                            <div class="col-2">
-                                <plex-datetime type="date" [(ngModel)]="fecha" (change)="refreshSelection($event,'fecha')" name="fecha" label="Fecha"
-                                    class="fechas" [required]="true">
-                                </plex-datetime>
-                            </div>
-
-                            <div class="col-4">
-                                <!-- Select Unidad Operativa -->
-                                <plex-select [(ngModel)]="unidadOperativa" name="unidadOperativa" [data]="opcionesUnidadOperativa"
-                                    label="Unidad Operativa" placeholder="Seleccione la unidad operativa" labelField="nombre"
-                                    (change)="refreshSelection($event,'unidadOperativa')" [required]="true" [readonly]="(organizacion === null)">
-                                </plex-select>
-                            </div>
-
-                        </div>
-
-                        <br>
-                    </header>
-
-                    <resumenDiarioMensual *ngIf="showResumenDiarioMensual" [reporte]="reporte" [parametros]="parametros" id="ResumenDiarioMensual"></resumenDiarioMensual>
-                    <planillaC1 *ngIf="showPlanillaC1" [reporte]="reporte" [parametros]="parametros" id="PlanillaC1"></planillaC1>
-
-                </plex-box>
-            </div>
-        </div>
-
-    </section>
-
-    <footer>
-        <plex-button type="success" label="Generar Reporte" (click)="generar()" class="float-right"></plex-button>
-        <plex-button *ngIf="showBotonImprimir" type="default" label="Imprimir Reporte" (click)="imprimir(parametros['tipoReportes'])" class="float-right"></plex-button>
-    </footer>
-
-</form>
